보스 개발 거의 완료
이제 던전 시스템 프로토타입 개발 시작

+++ Assets/02. Scripts/Enemy/EnemyModel.cs
@@ -156,7 +156,7 @@ public class EnemyModel : EnemyBase
 
     
 
-    public void Heal(float healAmount)
+    public void Heal(int healAmount)
     {
         if (_isDead) return;
